Fix syncronisation of vim commands with shell cmds
We now use neovim-remote to remotely tell vim to execute a command from
the shell. This is a bit convoluted but it means that we actually have
the correct execution order and syncronisation when a task mixes vim
commands and shell commands.

Also finally fix compile_and_run for a new CMake project. You no longer
have to run it twice!